Move trading params out of class to package.
